Textual conflict in tests/flow/test_null_handling.py was additive: main
extended test08_null_value_hash_join with the nested-null join coverage
from #2775, while this branch appended test09_null_simple_case after it.
Kept both.

Main also grew two things that pin the behaviour this branch changes, and
that git could not see as conflicts:

- graph/src/runtime/vector_expr.rs gained a columnar eval_case that
  matches the value form with `PartialEq`, exactly as the per-row one used
  to. Its own comment says the two must not diverge. Queries like
  `RETURN CASE null WHEN null ...` now take that path, so without this the
  branch's fix is a no-op and its own test09_null_simple_case fails on the
  merge. Switched it to compare_values(.., CmpOp::Eq), the shared helper
  that already spells out `=`'s truth table for the comparison kernels.

- tests/flow/test_filters.py's test09_case_value_form_null_subject
  asserted the old answer ('matched') both directly and as the
  columnar/per-row agreement check. Flipped the expectations to 'else';
  the agreement invariant the test exists to hold is unaffected.

Verified: flow suite 1510/1511 (the one failure is test_udf_cluster, a
3-shard cluster that fails to form locally before any query runs),
cargo test -p graph 403 passed, fmt and clippy clean.

+++ graph/src/index/vector_index_options.rs
@@ -1,6 +1,9 @@
-#[derive(Debug, Default, Clone)]
+#[derive(Debug, Default, Clone, PartialEq)]
 pub struct VectorIndexOptions {
-    pub dimension: u32,
+    /// `u64`, matching C's `size_t`, the v19 RDB's field and the effects
+    /// wire. A `u32` here meant every RDB load narrowed and every wire read
+    /// had to check.
+    pub dimension: u64,
     pub similarity_function: Option<String>,
     pub m: Option<usize>,
     pub ef_construction: Option<usize>,
